Got the map/maze working
The key insight is that you have to try every single direction
and by try, it means check if the new x and y position are valid
and also if there are blocks to remove. Another key insight is
that always assume that the block you are currently sitting on
has already been cleared.