Adding test case and fixed pushback with intList class
I am now testing out the IntList class and making sure that
everything is sound. Noticed that POPBACK (NOT pushback) was
not changing the tail pointer in the event that there was
only one node and both head and tail where pointing to
that same node. This would result in errors later on, but
now that case is fixed